+++ Moviebeam/app/src/main/java/com/diipl/moviebeam/ui/exoplayer/ExoPlayerActivity.kt
@@ -332,7 +332,7 @@ class ExoPlayerActivity : BaseActivity() {
                 request.productType = it.releaseTypeId
                 request.a = a
                 request.ra = 0
-                request.seekType = seekType
+                request.seekType = seekType.toLong()
                 request.seek = player.getLastSeek()
             }
             moviesViewModel.updateRentalMovieLog(request)
@@ -347,7 +347,7 @@ class ExoPlayerActivity : BaseActivity() {
                 request.productType = Constants.SHOWTIME_RELEASE_TYPE_ID
                 request.a = a
                 request.ra = 0
-                request.seekType = seekType
+                request.seekType = seekType.toLong()
                 request.seek = player.getLastSeek()
             }
             moviesViewModel.updateRentalMovieLog(request)
